version 1

+++ resources/views/pages/Animes/Anime.blade.php
@@ -2,10 +2,10 @@
 
 @section('content')
 
-    @if (Auth::user())
-        <section class="mb-5">
+    @if (Auth::user() && Auth::user()->role === 'admin')
+        <section class="mb-5 pt-5">
             {{-- btn --}}
-            <div class="container d-flex justify-content-between mt-5">
+            <div class="container d-flex justify-content-between">
                 <div><a href="{{ $anime->id }}/edit" class="btn btn-primary ">Edit</a></div>
                 <div><a href="" class="btn btn-info">Ajouter <br> un nouvelles Episode</a></div>
                 {!! Form::open(['action' => ['Animes\AnimesController@destroy', $anime->id], 'method' => 'DELETE']) !!}
@@ -22,7 +22,7 @@
     @endif
     
     {{-- Animes info display --}}
-    <section class="container d-flex flex-column mb-md-5">
+    <section class="container d-flex flex-column pt-5 mb-md-5">
         <div class=" d-flex flex-column flex-md-row">
 
             <div class="d-flex justify-content-center col-md-6">
@@ -30,6 +30,12 @@
             </div>
 
             <div class="d-flex flex-column col-md-8 p-3">
+               
+                    <div class="d-flex text-light">
+                        @foreach ($categories as $categorie)
+                            <span class="bg-info p-1 me-1">{{$categorie->nom}}</span>
+                        @endforeach
+                    </div>
                 <h1>{{ $anime->nom }}</h1>
                 <p>{{ $anime->synopsis }}</p>
             </div>
@@ -37,12 +43,15 @@
     </section>
 
     <section class="container-fluid d-flex row">
-        <h1 class="mb-3">Liste des episode</h1>
-        @foreach ($episodes as $episode)
-            <a href="/les_episodes/{{$episode->id}}/" class="d-flex flex-column col-5 col-md-3 mb-3 me-3 link-light">
-                <div class=""><img src="/storage/Animes/photos/{{ $anime->images }}" class="img-fluid" alt=""></div>
-                <h3>{{$episode->nom}}</h3>
-            </a>
-        @endforeach
+        <h1 class="mb-3">Liste des episodes</h1>
+
+        <div class="row mb-5 p-2">
+            @foreach ($anime->episodes as $episode)
+                <a href="/les_episodes/{{$episode->id}}" class="d-flex flex-column col-3 me-3 link-dark text-decoration-none hvr-shrink">
+                    <div class="responsive_bg_img" style="background-image: url(/storage/Animes/photos/{{ $anime->images }});height:25vh"></div>
+                    <div class="border shadow_box"><h5><b>{{$episode->nom}}</b></h5></div>
+                </a>
+            @endforeach
+        </div>
     </section>
 @endsection
